+++ server/src/main/java/org/opensearch/action/admin/indices/stats/IndicesStatsRequest.java
@@ -36,8 +36,12 @@ import org.opensearch.action.support.broadcast.BroadcastRequest;
 import org.opensearch.common.annotation.PublicApi;
 import org.opensearch.core.common.io.stream.StreamInput;
 import org.opensearch.core.common.io.stream.StreamOutput;
+import org.opensearch.core.tasks.TaskId;
+import org.opensearch.rest.action.admin.cluster.ClusterAdminTask;
+import org.opensearch.tasks.Task;
 
 import java.io.IOException;
+import java.util.Map;
 
 /**
  * A request to get indices level stats. Allow to enable different stats to be returned.
@@ -53,6 +57,8 @@ import java.io.IOException;
 public class IndicesStatsRequest extends BroadcastRequest<IndicesStatsRequest> {
 
     private CommonStatsFlags flags = new CommonStatsFlags();
+
+    private boolean isCancellationTaskRequired = false;
 
     public IndicesStatsRequest() {
         super((String[]) null);
@@ -101,6 +107,20 @@ public class IndicesStatsRequest extends BroadcastRequest<IndicesStatsRequest> {
     public IndicesStatsRequest groups(String... groups) {
         flags.groups(groups);
         return this;
+    }
+
+    public void setIsCancellationTaskRequired(boolean isCancellationTaskRequired) {
+        this.isCancellationTaskRequired = isCancellationTaskRequired;
+    }
+
+    @Override
+    public Task createTask(long id, String type, String action, TaskId parentTaskId, Map<String, String> headers) {
+        if (this.isCancellationTaskRequired) {
+            return new ClusterAdminTask(id, type, action, parentTaskId, headers);
+        }
+        else {
+            return super.createTask(id, type, action, parentTaskId, headers);
+        }
     }
 
     public String[] groups() {