+++ server/src/main/java/org/opensearch/action/support/TimeoutTaskCancellationUtility.java
@@ -12,11 +12,13 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.message.ParameterizedMessage;
 import org.opensearch.action.admin.cluster.node.tasks.cancel.CancelTasksRequest;
+import org.opensearch.action.admin.cluster.node.tasks.cancel.CancelTasksResponse;
 import org.opensearch.client.OriginSettingClient;
 import org.opensearch.client.node.NodeClient;
 import org.opensearch.common.settings.ClusterSettings;
 import org.opensearch.common.unit.TimeValue;
 import org.opensearch.core.action.ActionListener;
+import org.opensearch.core.tasks.TaskCancelledException;
 import org.opensearch.core.tasks.TaskId;
 import org.opensearch.search.SearchService;
 import org.opensearch.tasks.CancellableTask;
@@ -90,6 +92,70 @@ public class TimeoutTaskCancellationUtility {
                             )
                         )
                     );
+            });
+            wrappedListener.cancellable = client.threadPool().schedule(wrappedListener, timeoutInterval, ThreadPool.Names.GENERIC);
+            listenerToReturn = wrappedListener;
+        } catch (Exception ex) {
+            // if there is any exception in scheduling the cancellation task then continue without it
+            logger.warn("Failed to schedule the cancellation task for original task: {}, will continue without it", taskToCancel.getId());
+        }
+        return listenerToReturn;
+    }
+
+    /**
+     * Wraps and handle a listener with a timeout listener {@link TimeoutRunnableListener} to schedule the task cancellation for provided tasks on
+     * generic thread pool
+     * @param client - {@link NodeClient}
+     * @param taskToCancel - task to schedule cancellation for
+     * @param timeout - {@link TimeValue}
+     * @param listener - original listener associated with the task
+     * @return wrapped listener
+     */
+    public static <Response> ActionListener<Response> wrapAndHandleWithCancellationListener(
+        NodeClient client,
+        CancellableTask taskToCancel,
+        TimeValue timeout,
+        ActionListener<Response> listener
+    ) {
+        final TimeValue timeoutInterval = (taskToCancel.getCancellationTimeout() == null) ? timeout : taskToCancel.getCancellationTimeout();
+        // Note: -1 (or no timeout) will help to turn off cancellation. The combinations will be request level set at -1 or request level
+        // set to null and cluster level set to -1.
+        ActionListener<Response> listenerToReturn = listener;
+        if (timeoutInterval.equals(SearchService.NO_TIMEOUT)) {
+            return listenerToReturn;
+        }
+
+        try {
+            final TimeoutRunnableListener<Response> wrappedListener = new TimeoutRunnableListener<>(timeoutInterval, listener, () -> {
+                final CancelTasksRequest cancelTasksRequest = new CancelTasksRequest();
+                cancelTasksRequest.setTaskId(new TaskId(client.getLocalNodeId(), taskToCancel.getId()));
+                cancelTasksRequest.setReason("Cancellation timeout of " + timeoutInterval + " is expired");
+                // force the origin to execute the cancellation as a system user
+                new OriginSettingClient(client, TASKS_ORIGIN).admin()
+                    .cluster()
+                    .cancelTasks(cancelTasksRequest, new ActionListener<CancelTasksResponse>() {
+                        @Override
+                        public void onResponse(CancelTasksResponse cancelTasksResponse) {
+                            logger.debug(
+                                "Scheduled cancel task with timeout: {} for original task: {} is successfully completed",
+                                timeoutInterval,
+                                cancelTasksRequest.getTaskId()
+                            );
+                            listener.onFailure(new TaskCancelledException(cancelTasksRequest.getReason()));
+                        }
+
+                        @Override
+                        public void onFailure(Exception e) {
+                            logger.error(
+                                new ParameterizedMessage(
+                                    "Scheduled cancel task with timeout: {} for original task: {} is failed",
+                                    timeoutInterval,
+                                    cancelTasksRequest.getTaskId()
+                                ),
+                                e
+                            );
+                        }
+                    });
             });
             wrappedListener.cancellable = client.threadPool().schedule(wrappedListener, timeoutInterval, ThreadPool.Names.GENERIC);
             listenerToReturn = wrappedListener;